moving the hypermixing module to the nnet folder and refactoring dependencies
