+++ speechbrain/lobes/models/transformer/Branchformer.py
@@ -17,7 +17,7 @@ from speechbrain.nnet.attention import RelPosMHAXL, MultiheadAttention
 from speechbrain.nnet.normalization import LayerNorm
 from speechbrain.lobes.models.convolution import ConvolutionalSpatialGatingUnit
 
-from speechbrain.lobes.models.transformer.hypermixing import HyperMixing
+from speechbrain.nnet.hypermixing import HyperMixing
 
 
 class ConvolutionBranch(nn.Module):
